+++ Vector.h
@@ -9,14 +9,14 @@ class Vector : public Point
 {
 public:
 	Vector();
-	Vector(int, int, int);
+	Vector(double, double, double);
 	Vector(const Point&, const Point&);
 	Vector(const Vector&);
 	Vector& operator=(const Vector&);
 	~Vector();
 
 	//task 2
-	int vectorLength();
+	double vectorLength();
 	Vector vectorDirection();
 	bool isVectorNull();
 	bool vectorParallelism(Vector&);
@@ -26,8 +26,8 @@ public:
 
 	Vector operator +(const Vector&);
 	Vector operator -(const Vector&);
-	int operator *(const Vector&);
-	int operator()(const Vector&, const Vector&);
+	double operator *(const Vector&);
+	double operator()(const Vector&, const Vector&);
 
 	virtual std::ostream& ins(std::ostream&)const;
 	virtual std::istream& ext(std::istream&);
